sample-scripts/requirements.txt
python-frontmatter==0.5.0
pyyaml==5.4.1